LLM_Academic_Advisor/AcademicAdvisorBot/requirements.txt
streamlit
langchain
openai
faiss-cpu
PyPDF2
python-dotenv
tiktoken
